GUI já apresenta o JSON devidamente identado e as propriedades no lado do editor (WIP)
