+++ src/main/kotlin/model/JSONMap.kt
@@ -30,14 +30,14 @@ class JSONMap(map: Map<*, *>) : JSONElement {
     }
 
     override fun toString(): String {
-        val str = StringBuilder().append("{")
+        val str = StringBuilder().append("{\n")
         value.entries.forEach {
             if(value.keys.indexOf(it.key) != value.size - 1){
-                str.append("\"" + it.key + "\": " + it.value.toString() + ", ")
+                str.append(("\"" + it.key + "\": " + it.value.toString() + ",").prependIndent("\t") + "\n")
             }else{
-                str.append("\"" + it.key + "\": " + it.value.toString())
+                str.append(("\"" + it.key + "\": " + it.value.toString()).prependIndent("\t"))
             }
         }
-        return str.append("}").toString()
+        return str.append("\n}").toString()
     }
 }
