+++ src/main/kotlin/model/UC.kt
@@ -1,3 +1,3 @@
 package model
 
-data class UC(val uc: String, val ects: Double, val dataExame: String, val inscritos: List<Aluno>){}
+data class UC(val uc: String, val ects: Double, val dataExame: String, val inscritos: List<Aluno>, val cursos: List<String>){}
